feat: show replace directory info
