+++ src/main.ts
@@ -3,7 +3,7 @@ import { setFailed, getInput } from '@actions/core';
 import { context, GitHub } from '@actions/github';
 import signale from 'signale';
 import { deploy } from './utils/command';
-import { getBuildVersion, isTargetEvent, getTagName, isValidTagName } from './utils/misc';
+import { getBuildVersion, isTargetEvent, getTagName, isValidTagName, getReplaceDirectory } from './utils/misc';
 
 /**
  * run
@@ -23,6 +23,9 @@ async function run(): Promise<void> {
 			signale.complete('This is not target event.');
 			return;
 		}
+
+		const directories = getReplaceDirectory();
+		Object.keys(directories).forEach(directory => signale.info('%s: %s', directories[directory], directory));
 
 		await deploy(new GitHub(getInput('GITHUB_TOKEN', {required: true})), context);
 	} catch (error) {
